+++ R/check_model.fit_model_napping.R
@@ -1,0 +1,5 @@
+check_model.fit_model_napping = function(x){
+  out = x
+  class(out) <- c("PPBstats", "check_model_napping", "MFA")
+  return(out) 
+}