+++ R/common_functions.R
@@ -740,8 +740,6 @@ plot_check_freq_anova = function(x){
   data_ggplot_variability_repartition_pie = data_ggplot$data_ggplot_variability_repartition_pie
   data_ggplot_var_intra = data_ggplot$data_ggplot_var_intra
   
-  data_ggplot_pca = x$GxE$PCA
-  
   # 1. Normality ----------
   # 1.1. Histogram ----------
   p = ggplot(data_ggplot_normality, aes(x = r), binwidth = 2)
@@ -773,18 +771,13 @@ plot_check_freq_anova = function(x){
   p = p + theme(legend.position = "none", axis.text.x = element_text(angle = 90), plot.title=element_text(hjust=0.5))
   p3 = p 
   
-  # 4. pca composante variance ----------
-  p4 = fviz_eig(data_ggplot_pca) + ggtitle("")
-  
-  # 5. return results
-  
+  # 4. return results
   out = list(
     "residuals" = list(
       "histogram" = p1.1,
       "qqplot" = p1.2),
     "variability_repartition" = p2,
-    "variance_intra_germplasm" = p3,
-    "pca_composante_variance" = p4
+    "variance_intra_germplasm" = p3
   )
   
   return(out)
@@ -1084,5 +1077,64 @@ add_pies = function(p, n, format, plot_type, data_to_pie, variable, pie_size){
   }
   
   return(p)
+}
+
+# format_organo ----------
+format_organo = function(data, threshold){
+  # 1. Merge and create data frame ----------
+  N = data
+  N$sample = factor(paste(N$location, N$germplasm, sep = ":"))
+  
+  # 2. Add the occurence of the different descriptors ----------
+  descriptors = as.vector(as.character(N$descriptors))
+  vec_adj = unlist(strsplit(descriptors, ";"))
+  vec_adj = sort(unique(vec_adj))
+  if( length(which(vec_adj == "")) > 0 ) { vec_adj = vec_adj[-which(vec_adj == "")] }
+  
+  df = matrix(0, ncol = length(vec_adj), nrow = nrow(N))
+  df = as.data.frame(df)
+  colnames(df) = vec_adj
+  out = cbind.data.frame(N, df)
+  
+  for (i in 1:nrow(out)){
+    v_adj = out[i, "descriptors"]
+    v_adj = unlist(strsplit(as.character(v_adj), ";"))
+    if( length(which(v_adj == "")) > 0 ) { v_adj = v_adj[-which(v_adj == "")] }
+    
+    for (j in 1:length(v_adj)) {
+      e = v_adj[j]
+      if (length(e)>0) {
+        if (!is.na(e)) { out[i, e] = 1 }
+      }
+    }
+  }
+  
+  N = out[,-which(colnames(out) == "descriptors")]
+  
+  # 3. Apply the threshold to keep certain descriptors ----------
+  if( !is.null(threshold) ) {
+    test = apply(N[, vec_adj], 2, sum)
+    to_delete = which(test <= threshold)
+    to_keep = which(test > threshold)
+    if( length(to_delete) > 0 ) { 
+      adj_to_delete = vec_adj[to_delete] 
+      N = N[,-which(colnames(N) %in% adj_to_delete)]
+      message("The following descriptors have been remove because there were less or equal to ", threshold, " occurences : ", paste(adj_to_delete, collapse = ", "))
+      if( ncol(N) == 4 ){ stop("There are no more descriptors with threshold = ", threshold, 
+                               ". You must set another value.") }
+    }
+    vec_adj = vec_adj[to_keep]
+  }
+  
+  
+  # 4. Get frequency for each descriptor ----------
+  N_freq = N_raw = N
+  for (ad in vec_adj) { 
+    if( sum(N_raw[, ad], na.rm = TRUE) != 0 ) { 
+      N_freq[, ad] = N_raw[, ad] / sum(N_raw[, ad], na.rm = TRUE) 
+    }  
+  }
+  
+  return(N_freq)
 }