+++ R/plot.check_model_napping.R
@@ -1,0 +1,6 @@
+plot.check_model_napping <- function(
+  x
+){
+  out = fviz_eig(x) + ggtitle("")
+  return(out)
+}